console: Add useDerivedRightsProp hook
